requirements.txt
# requirements.txt
streamlit==1.28.1
pandas==2.0.3
openpyxl==3.1.2
langchain-groq==0.1.1
langgraph==0.0.42
pydantic==2.5.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-json-report==1.5.0